Don't start Spotify when checking if it's running
`Application("Spotify").running()` will start Spotify (!) then check if
it's running.

To work around this and check if Spotify is running _without_ starting
it, use System Events and see if `Spotify` is in the names of running
applications.

Now I can use a fast refresh in tmux without the 1-second refresh
restarting Spotify before Spotify has a chance to shut down!

Also, sometimes when Spotify just started, `track` is `null`, so add a
quick check around that.

More info here: https://stackoverflow.com/questions/16064957/how-to-check-in-applescript-if-an-app-is-running-without-launching-it-via-osa